defi-analytics-pipeline/extract_load/requirements.txt
requests>=2.31.0
pandas>=2.0.0
web3>=6.0.0
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0